# swagger/swagger.json
{"swagger":"2.0","info":{"version":""},"consumes":["application/json","application/xml","application/gob","application/x-gob"],"produces":["application/json","application/xml","application/gob","application/x-gob"],"paths":{"/auth/authorize-client":{"get":{"tags":["authUI"],"summary":"promptAuthorization authUI","description":"Prompt the user for client authorization","operationId":"authUI#promptAuthorization","produces":["application/vnd.goa.error","application/vnd.goa.error"],"responses":{"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/error"}}}}},"/auth/confirm-authorization":{"post":{"tags":["authUI"],"summary":"confirmAuthorization authUI","description":"Confirm the authorization of the client","operationId":"authUI#confirmAuthorization","produces":["application/vnd.goa.error","application/vnd.goa.error"],"parameters":[{"name":"confirmed","in":"query","description":"Is the authorization confirmed.","required":false,"type":"boolean"}],"responses":{"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/error"}}}}},"/js/*.js":{"get":{"summary":"Download public/js/*.js","operationId":"public#/js/*.js","responses":{"200":{"description":"File downloaded","schema":{"type":"file"}}}}},"/login":{"get":{"tags":["login"],"summary":"showLogin login","description":"Shows a login screen","operationId":"login#showLogin","produces":["application/vnd.goa.error","application/vnd.goa.error","application/vnd.goa.error"],"responses":{"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"401":{"description":"Unauthorized","schema":{"$ref":"#/definitions/error"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/error"}}}}},"/oauth2/authorize":{"get":{"tags":["oauth2_provider"],"summary":"authorize oauth2_provider","description":"Authorize OAuth2 client","operationId":"oauth2_provider#authorize","produces":["application/vnd.goa.example.oauth2.error+json"],"parameters":[{"name":"client_id","in":"query","description":"The client identifier","required":true,"type":"string"},{"name":"redirect_uri","in":"query","description":"Redirection endpoint","required":false,"type":"string"},{"name":"response_type","in":"query","description":"Value MUST be set to \"code\"","required":true,"type":"string","enum":["code"]},{"name":"scope","in":"query","description":"The scope of the access request","required":false,"type":"string"},{"name":"state","in":"query","description":"An opaque value used by the client to maintain state between the request and callback","required":false,"type":"string"}],"responses":{"302":{"description":"Found","headers":{"Location":{"description":"Redirect URL containing the authorization code and state param if any","type":"string"}}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/OAuth2ErrorMedia"}}}}},"/oauth2/token":{"post":{"tags":["oauth2_provider"],"summary":"get_token oauth2_provider","description":"Get access token from authorization code or refresh token","operationId":"oauth2_provider#get_token","produces":["application/vnd.goa.example.oauth2.error+json","application/vnd.goa.example.oauth2.token+json"],"parameters":[{"name":"payload","in":"body","description":"Payload sent by client to obtain refresh and access token or to refresh an access token.\nsee https://tools.ietf.org/html/rfc6749#section-4.1.3 and https://tools.ietf.org/html/rfc6749#section-6","required":true,"schema":{"$ref":"#/definitions/TokenPayload"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/TokenMedia"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/OAuth2ErrorMedia"}}},"security":[{"oauth2_client_basic_auth":[]}]}}},"definitions":{"OAuth2ErrorMedia":{"title":"Mediatype identifier: application/vnd.goa.example.oauth2.error+json; view=default","type":"object","properties":{"error":{"type":"string","description":"Error returned by authorization server","example":"invalid_request","enum":["invalid_request","invalid_client","invalid_grant","unauthorized_client","unsupported_grant_type"]},"error_description":{"type":"string","description":"Human readable ASCII text providing additional information","example":"Omnis est ex sed."},"error_uri":{"type":"string","description":"A URI identifying a human-readable web page with information about the error","example":"Et eos consequatur autem deleniti sed."}},"description":"OAuth2 error response, see https://tools.ietf.org/html/rfc6749#section-5.2 (default view)","example":{"error":"invalid_request","error_description":"Omnis est ex sed.","error_uri":"Et eos consequatur autem deleniti sed."},"required":["error"]},"TokenMedia":{"title":"Mediatype identifier: application/vnd.goa.example.oauth2.token+json; view=default","type":"object","properties":{"access_token":{"type":"string","description":"The access token issued by the authorization server","example":"Nisi ex velit amet eius tempora iusto."},"expires_in":{"type":"integer","description":"The lifetime in seconds of the access token","example":1657379562044974888,"format":"int64"},"refresh_token":{"type":"string","description":"The refresh token","example":"Numquam velit et explicabo id."},"scope":{"type":"string","description":"The scope of the access token","example":"Qui dignissimos corrupti similique voluptatibus a alias."},"token_type":{"type":"string","description":"The type of the token issued, e.g. \"bearer\" or \"mac\"","example":"Est minima vel possimus enim saepe."}},"description":"OAuth2 access token request successful response, see https://tools.ietf.org/html/rfc6749#section-5.1 (default view)","example":{"access_token":"Nisi ex velit amet eius tempora iusto.","expires_in":1657379562044974888,"refresh_token":"Numquam velit et explicabo id.","scope":"Qui dignissimos corrupti similique voluptatibus a alias.","token_type":"Est minima vel possimus enim saepe."},"required":["access_token","token_type"]},"TokenPayload":{"title":"TokenPayload","type":"object","properties":{"code":{"type":"string","description":"The authorization code received from the authorization server, used for initial refresh and access token request","example":"Delectus aut libero nesciunt veniam reiciendis."},"grant_type":{"type":"string","description":"Value MUST be set to \"authorization_code\" when obtaining initial refresh and access token.\nValue MUST be set to \"refresh_token\" when refreshing an access token.","example":"refresh_token","enum":["authorization_code","refresh_token"]},"redirect_uri":{"type":"string","description":"The redirect_uri parameter specified when making the authorize request to obtain the authorization code, used for initial refresh and access token request","example":"Ipsum et aut."},"refresh_token":{"type":"string","description":"The refresh token issued to the client, used for refreshing an access token","example":"Et qui quia sed odio sint."},"scope":{"type":"string","description":"The scope of the access request, used for refreshing an access token","example":"Iste laborum nostrum quidem sequi dolor."}},"description":"Payload sent by client to obtain refresh and access token or to refresh an access token.\nsee https://tools.ietf.org/html/rfc6749#section-4.1.3 and https://tools.ietf.org/html/rfc6749#section-6","example":{"code":"Delectus aut libero nesciunt veniam reiciendis.","grant_type":"refresh_token","redirect_uri":"Ipsum et aut.","refresh_token":"Et qui quia sed odio sint.","scope":"Iste laborum nostrum quidem sequi dolor."},"required":["grant_type"]},"error":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"code":{"type":"string","description":"an application-specific error code, expressed as a string value.","example":"invalid_value"},"detail":{"type":"string","description":"a human-readable explanation specific to this occurrence of the problem.","example":"Value of ID must be an integer"},"id":{"type":"string","description":"a unique identifier for this particular occurrence of the problem.","example":"3F1FKVRR"},"meta":{"type":"object","description":"a meta object containing non-standard meta-information about the error.","example":{"timestamp":1458609066},"additionalProperties":true},"status":{"type":"string","description":"the HTTP status code applicable to this problem, expressed as a string value.","example":"400"}},"description":"Error response media type (default view)","example":{"code":"invalid_value","detail":"Value of ID must be an integer","id":"3F1FKVRR","meta":{"timestamp":1458609066},"status":"400"}}},"securityDefinitions":{"OAuth2":{"type":"oauth2","flow":"accessCode","authorizationUrl":"/oauth2/authorize","tokenUrl":"/oauth2/token","scopes":{"api:read":"no description","api:write":"no description"}},"oauth2_client_basic_auth":{"type":"basic","description":"Basic auth used by client to make the requests needed to retrieve and refresh access tokens"}}}